requirements.txt
python-telegram-bot[webhooks]==20.3
web3==6.4.0
mnemonic==0.20
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.2
gunicorn==20.1.0
bip-utils==2.7.0
pytonlib==0.0.42
prettytable==3.8.0
solana==0.29.0
spl-token==3.2.0
sqlalchemy==2.0.19
apscheduler==3.10.1